feat: Add float ranges support (#109)
* fix: generators emit wrong numeric type for floating-point ranges (#78)

cddl2java, cddl2py, cddl2swift and cddl2kotlin all rendered a range like
(0.0..1.0) using the integer type, indistinguishable from (0..1), because
whole-valued float bounds collapse to integers under Number.isInteger.
Parser PR #77 added an IsFloat marker to numeric literal nodes to make the
distinction possible again; this wires that marker into all four
generators' range handling.

Adds a shared isFloatRange(range) helper to the cddl package rather than
duplicating the check per generator. It also handles Min/Max as a plain
number (the Range type's declared shape) in addition to the literal-node
object the parser actually emits, since at least one existing test double
(cddl2py's rangeRef()) already relied on the plain-number shape.

Along the way, fixed two related bugs in the same code paths beyond what
the issue's example covered:
- cddl2java always emitted Float for any range with an operator (e.g. a
  default value), even a genuinely integer range like `twist: (0..359)
  .default 0`-style declarations - now correctly Integer.
- cddl2py/cddl2swift/cddl2kotlin's generateVariable() had its own,
  separate hardcoded-int/Int/Long branch for a bare `MyRange = (0..1)`
  style variable, not just the resolveType() branch the issue quoted.

Verified against the real webdriver CDDL spec: pressure, tangentialPressure,
altitudeAngle, azimuthAngle and quality (all genuinely 0.0..1.0-style
ranges) now correctly render as Float/float/Double, while twist (0..359,
an integer range that happened to carry a default operator) now correctly
renders as Integer instead of the previous always-Float bug.

cddl2ts needs no change - TypeScript has no separate int/float numeric
type, so `number` was already correct.


* docs+test: close doc/coverage gaps left by the float-range fix (#78)

Docs:
- packages/cddl/docs/references.md: the canonical PropertyReference type
  block was missing the IsFloat field entirely (added by parser PR #77,
  before this session's work even started) - add it with an explanation.
- packages/cddl/docs/ranges.md: the "Range with Floating-Point Values"
  and "Ranges with Operators" sections' AST examples predated IsFloat and
  showed Min/Max as bare numbers, which can't represent it at all. Fixed
  both to show the real literal-node shape (verified by running the
  parser on the doc's own `scale = (0.1..2) .default 1` example) and
  explain why only one bound needs IsFloat for the range to be float.

Test coverage, found via branch-level (not just line-level) coverage
inspection of every isFloatRange() call site added in the previous commit:

- cddl2java: exported parseType() (previously module-private) so it can
  be unit tested directly - the package's only prior tests are
  file-based end-to-end fixtures against a BiDi-spec-shaped bidi.cddl,
  which silently drops any top-level definition that doesn't match BiDi's
  module/command/result naming convention, so a small synthetic .cddl
  fixture wasn't viable. Added packages/cddl2java/tests/parseType.test.ts
  covering both the with-operator and without-operator range branches,
  each with both a float and a genuine integer range - the latter had
  zero coverage in the existing bidi.cddl fixture for the no-operator
  branch specifically.
- cddl2swift/cddl2kotlin: generateVariable()'s bare range-to-typealias
  branch (a top-level `MyRange = (0..1)` style variable) had never been
  exercised by any test, int or float. resolveType()'s two range branches
  were each only ever exercised with one outcome by the real webdriver
  spec (e.g. the operator-wrapped branch only ever saw float, via
  `scale`; the operator-less branch only ever saw integer). Added 3 tests
  per package covering every previously-missing combination.

Verified via packages/cddl/tests/utils.test.ts + coverage-final.json
branch inspection that every isFloatRange() call site's ternary now has
both outcomes exercised by at least one test, not just the statement
itself being reached. Full checks:all passes.


* docs: explain the RFC 8610 BAD-range rationale in isFloatRange's JSDoc

isFloatRange() picking float when only one bound is float was already
correct and tested, but the JSDoc didn't say why: RFC 8610 defines only
int-range (both bounds int) and float-range (both bounds float) - a mixed
range is explicitly "NOT DEFINED" by the spec. Document that we pick
float for that undefined case since it's the non-lossy choice.


* fix: isFloatRange misclassified symbolic range bounds as float

isFloatEndpoint's object branch ran `!Number.isInteger(node.Value)`
unconditionally. A symbolic/named bound - e.g. RFC 8610's own
`max-byte = 255` / `byte = 0..max-byte` example - parses as
`{ Type: 'group', Value: 'max-byte' }`, a string Value.
Number.isInteger('max-byte') is false, so the unguarded check made any
range with a named upper/lower bound look like a float range, regardless
of what the reference actually resolves to.

Guard with `typeof node.Value === 'number'` before the integer check, so
a non-numeric Value (string reference, or anything else) can never force
a range to float on its own - only an actual float literal (via IsFloat
or a genuinely non-integer number) can.

Reproduced with the RFC's own byte/byte1 example before fixing.


---------

+++ packages/cddl2java/tests/__fixtures__/Input/PointerCommonProperties.java
@@ -23,7 +23,7 @@ public class PointerCommonProperties {
     /**
      * Creates a new PointerCommonProperties instance
      */
-    public PointerCommonProperties(Long width, Long height, Float pressure, Float tangentialPressure, Float twist, Float altitudeAngle, Float azimuthAngle) {
+    public PointerCommonProperties(Long width, Long height, Float pressure, Float tangentialPressure, Integer twist, Float altitudeAngle, Float azimuthAngle) {
         this.width = width;
         this.height = height;
         this.pressure = pressure;
@@ -73,13 +73,13 @@ public class PointerCommonProperties {
         return this.tangentialPressure;
     }
 
-    private final Float twist;
+    private final Integer twist;
 
     /**
      * Gets the twist property
-     * @return Float value
+     * @return Integer value
      */
-    public Float getTwist() {
+    public Integer getTwist() {
         return this.twist;
     }
